add delay to skills bar function
Delay should help so that when a viewer first clicks on the About Me page, a 1.5 second delay begins to allow time to start skimming and scrolling down the page. Delay is not so long as to be extremely noticeable if user refreshes while at the bottom of the page.